// src/interfaces/CotizacionInterface.tsx
export interface getCollectionParams {
  currentPage: number,
  rowsPerPage: number,
  reload?: boolean
}

// Generated by https://quicktype.io

export interface Cotizacion {
  datos:         Dato[];
  desde:         number;
  hasta:         number;
  por_pagina:    number;
  pagina_actual: number;
  ultima_pagina: number;
  total:         number;
}

export interface Dato {
  id:                          number;
  numero_cotizacion_servicio:  number;
  solicitud_cotizacion_id:     number;
  empresa_cotizacion:          string;
  asociado_id:                 number;
  fecha_cotizacion:            string;
  fecha_vigencia_cotizacion:   string;
  plazo_pago_cotizacion:       number;
  numero_viajes_mes:           number;
  observaciones:               null | string;
  estado_cotizacion:           string;
  estado:                      number;
  usuario_creacion_id:         number;
  usuario_creacion_nombre:     string;
  usuario_modificacion_id:     number;
  usuario_modificacion_nombre: string;
  fecha_creacion:              string;
  fecha_modificacion:          string;
  numero_solicitud:            number;
  fecha_solicitud_cotizacion:  string;
  numero_documento:            string;
  tipo_cotizacion:             string;
  nombre_empresa:              string;
  nombre_asociado:             string;
}

// Generated by https://quicktype.io

export interface CotizacionShow {
  id:                          number;
  numero_cotizacion_servicio:  number;
  solicitud_cotizacion_id:     number;
  numero_solicitud_cotizacion: number;
  empresa_cotizacion:          string;
  asociado_id:                 number;
  numero_servicios_mes:        number|null;
  fecha_cotizacion:            string;
  fecha_vigencia_cotizacion:   string;
  plazo_pago_cotizacion:       number;
  numero_viajes_mes:           number;
  observaciones:               string|null;
  estado_cotizacion:           string;
  estado:                      number;
  usuario_creacion_id:         number;
  usuario_creacion_nombre:     string;
  usuario_modificacion_id:     number;
  usuario_modificacion_nombre: string;
  fecha_creacion:              string;
  fecha_modificacion:          string;
  solicitud_cotizacion:        SolicitudCotizacion;
  nombre_empresa:              string;
  asociado:                    Asociado;
}

export interface Asociado {
  id:     number;
  nombre: string;
}

export interface SolicitudCotizacion {
  id:                         number;
  numero_solicitud:           number;
  fecha_solicitud_cotizacion: string;
}

// Generated by https://quicktype.io

export interface AsociadoNegocio {
  id:               number;
  nombre:           string;
  numero_documento: string;
  estado:           number;
}